BT A2DP AAC encoder lacks sufficient calculation precision
Bug:515322464

1. AAC frame_length is 4096 bytes, sample_rate is 44100 Hz,
than the interval for each data acquisition
(a2dp_aac_encoder_interval_ms` is 23 ms.
2. bytes_per_tick represents the amount of audio data
required for each cycle (23 ms). The actual value
should be 4057.2 bytes. however, because bytes_per_tick
uses the integer data type uint32_t, it results in
bytes_per_tick being truncated to 4057.

Define counter and bytes_per_tick as double types.

+++ system/stack/a2dp/a2dp_aac_encoder.cc
@@ -64,8 +64,8 @@ typedef struct {
 } tA2DP_AAC_ENCODER_PARAMS;
 
 typedef struct {
-  float counter;
-  uint32_t bytes_per_tick; /* pcm bytes read each media task tick */
+  double counter;
+  double bytes_per_tick; /* pcm bytes read each media task tick */
   uint64_t last_frame_us;
 } tA2DP_AAC_FEEDING_STATE;
 
@@ -416,10 +416,10 @@ void a2dp_aac_feeding_reset(void) {
   memset(&a2dp_aac_encoder_cb.aac_feeding_state, 0, sizeof(a2dp_aac_encoder_cb.aac_feeding_state));
 
   a2dp_aac_encoder_cb.aac_feeding_state.bytes_per_tick =
-          (a2dp_aac_encoder_cb.feeding_params.sample_rate *
+          ((double)a2dp_aac_encoder_cb.feeding_params.sample_rate *
            a2dp_aac_encoder_cb.feeding_params.bits_per_sample / 8 *
            a2dp_aac_encoder_cb.feeding_params.channel_count * a2dp_aac_encoder_interval_ms) /
-          1000;
+          1000.0;
 
   log::info("PCM bytes {} per tick {} ms", a2dp_aac_encoder_cb.aac_feeding_state.bytes_per_tick,
             a2dp_aac_encoder_interval_ms);
@@ -469,8 +469,8 @@ static void a2dp_aac_get_num_frame_iteration(uint8_t* num_of_iterations, uint8_t
   a2dp_aac_encoder_cb.aac_feeding_state.last_frame_us = now_us;
 
   a2dp_aac_encoder_cb.aac_feeding_state.counter +=
-          (float)a2dp_aac_encoder_cb.aac_feeding_state.bytes_per_tick * us_this_tick /
-          (a2dp_aac_encoder_interval_ms * 1000);
+          a2dp_aac_encoder_cb.aac_feeding_state.bytes_per_tick * (double)us_this_tick /
+          (a2dp_aac_encoder_interval_ms * 1000.0);
 
   result = a2dp_aac_encoder_cb.aac_feeding_state.counter / pcm_bytes_per_frame;
   a2dp_aac_encoder_cb.aac_feeding_state.counter -= result * pcm_bytes_per_frame;
